Rename messages -> lmp and move to src/
